Add files via upload
In this project, I've developed an IPv6 subnetting calculator that caters to the needs of a network engineer, like myself, who is also striving for IPv6 certification. Crafting tools tailored to the exigencies of our daily operations has become imperative. Gratitude is owed to dynamic programming languages like Python, which empower us to materialize these solutions.

Beyond its technical prowess, this program boasts versatility. It's deployable on your personal computing device, leveraging the Tkinter library to forge a visually engaging interface complete with input fields catering to IPv6 network specifics, the prefix, and the quantum of desired subnets.

Execution is a breeze. Merely replicate the provided code within a Python (.py) file and initiate the runtime. A prerequisite is the presence of Tkinter within your Python environment.